requirements.txt
streamlit
ultralytics
opencv-python-headless
matplotlib
numpy
pillow
